science2go-main/requirements.txt
# requirements.txt - Science2Go dependencies
# Install with: pip install -r requirements.txt
# Recommended: use conda environment (see environment.yml)

# PDF Processing
PyPDF2>=3.0.1

# Google Cloud Services
google-cloud-texttospeech>=2.27.0
google-generativeai>=0.8.0

# Audio Processing
pydub>=0.25.1
mutagen>=1.47.0

# Configuration & Templates
PyYAML>=6.0.1
python-dotenv>=1.0.0

# Utilities
requests>=2.31.0
platformdirs>=4.0.0

# Modern GUI
customtkinter>=5.2.0

# PDF to Markdown conversion (optional - pulls in PyTorch, ~2GB)
# Uncomment to enable PDF-to-Markdown feature:
# marker-pdf>=1.10.0
# pdftext>=0.6.0
